Refactor proteus-registry
It has been aligned to follow the same directory structure and style of
proteus-orchestrate
