+++ proteus-registry/registry/handler/registry_test.go
@@ -1,4 +1,4 @@
-package registry
+package handler
 
 import (
 	"database/sql"
