+++ proteus-registry/registry/info.go
@@ -1,0 +1,4 @@
+package registry
+
+// LongDescription for the this package
+var LongDescription = `Is responsible for registering new clients and receiving updates from them`
